Added ability to "Edit On..." other branches.
Previously, the "Edit On..." link in the footer only allowed you to edit the current document in the master branch. This functionality allows a
user to set the repository branch they'd like the links to point to via an extra setting in the mkdocs yaml; not specifying a setting will
caust the link to default to master.

Added a page describing optional configurations to clarify this for users.